module-iface was extracted into standalone library
