+++ ydb/library/yql/providers/dq/provider/exec/yql_dq_exectransformer.cpp
@@ -1,10 +1,11 @@
 #include "yql_dq_exectransformer.h"
 
+#include <ydb/library/yql/module/yql_modules.h>
+
 #include <ydb/library/yql/providers/dq/provider/yql_dq_datasource.h>
 #include <ydb/library/yql/providers/dq/provider/yql_dq_state.h>
 
 #include <ydb/library/yql/providers/common/provider/yql_data_provider_impl.h>
-#include <ydb/library/yql/providers/common/provider/yql_modules.h>
 #include <ydb/library/yql/providers/common/provider/yql_provider.h>
 #include <ydb/library/yql/providers/common/provider/yql_provider_names.h>
 #include <ydb/library/yql/providers/common/transform/yql_exec.h>
