+++ ydb/library/yql/providers/yt/gateway/native/ya.make
@@ -26,6 +26,7 @@ PEERDIR(
     ydb/library/yql/ast
     ydb/library/yql/core/file_storage
     ydb/library/yql/minikql/comp_nodes
+    ydb/library/yql/module
     ydb/library/yql/utils
     ydb/library/yql/utils/log
     ydb/library/yql/utils/threading
