+++ ydb/library/yql/providers/yt/gateway/native/yql_yt_transform.cpp
@@ -1,7 +1,7 @@
 #include "yql_yt_transform.h"
 
+
 #include <ydb/library/yql/providers/yt/lib/skiff/yql_skiff_schema.h>
-#include <ydb/library/yql/providers/common/provider/yql_modules.h>
 #include <ydb/library/yql/providers/yt/common/yql_names.h>
 #include <ydb/library/yql/providers/yt/common/yql_configuration.h>
 #include <ydb/library/yql/providers/yt/codec/yt_codec.h>
@@ -9,6 +9,7 @@
 #include <ydb/library/yql/providers/yt/expr_nodes/yql_yt_expr_nodes.h>
 #include <ydb/library/yql/providers/common/codec/yql_codec_type_flags.h>
 
+#include <ydb/library/yql/module/yql_modules.h>
 #include <ydb/library/yql/utils/log/log.h>
 #include <ydb/library/yql/minikql/mkql_node_cast.h>
 #include <ydb/library/yql/utils/yql_panic.h>
